+++ bentoboxapp2/package.json
@@ -16,7 +16,9 @@
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-native": "0.73.5",
-    "expo-linear-gradient": "~12.7.2"
+    "expo-linear-gradient": "~12.7.2",
+    
+    "@react-native-async-storage/async-storage": "1.21.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
